separate consensus committee and decryption keyset
Still only supports a fixed decryption keyset.
Multi-keyset and sunset timestamps are not introduced.

The decryption keyset is hashed into the symmetric key
in the [SG01] scheme as metadata for additional in security.

+++ timeboost/src/binaries/keygen.rs
@@ -42,7 +42,7 @@ impl Scheme {
                 }
             }
             Self::Decryption => {
-                let (pub_key, comb_key, key_shares) = DecryptionScheme::trusted_keygen(num as u64);
+                let (pub_key, comb_key, key_shares) = DecryptionScheme::trusted_keygen(num as u16);
                 debug!("generating new threshold encryption keyset");
                 let pub_key = bs58_encode(&pub_key.as_bytes());
                 let comb_key = bs58_encode(&comb_key.as_bytes());
